Bug#27419181 UBSAN: MY_STRNTOLL_MB2_OR_MB4 - NAUGHTY NEGATIVES
Bad handling of -9223372036854775808, it cannot safely be negated.

+++ strings/ctype-ucs2.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2003, 2017, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2003, 2018, Oracle and/or its affiliates. All rights reserved.
    
    This library is free software; you can redistribute it and/or
    modify it under the terms of the GNU Library General Public
@@ -421,7 +421,7 @@ bs:
     return negative ? LLONG_MIN : LLONG_MAX;
   }
   
-  return (negative ? -((longlong)res) : (longlong)res);
+  return negative ? -res : res;
 }
